refactor naming in NP and revert abstraction for shadowComponent
